mcp-integrations/requirements.txt
## Requires Python 3.10+
mcp>=1.0.0
openai>=1.60.0
google-api-python-client>=2.155.0
google-auth>=2.36.0
google-auth-oauthlib>=1.2.1
notion-client>=2.2.0
python-dotenv>=1.0.0
